+++ src/constants.js
@@ -1,0 +1,28 @@
+const CONSTANTS = {
+    PLAN: 12,
+    WEEKEND_PLAN: 35,
+    PLAN_COEF: {
+        10: 1,
+        20: 1.65,
+        30: 2.7,
+        60: 4.5,
+        720: 7.5
+    },
+    BONUS: {
+        10: 60,
+        20: 85,
+        30: 162,
+        60: 245,
+        720: 400
+    },
+    WEEKEND_BONUS: {
+        10: 55,
+        20: 65,
+        30: 148.5,
+        60: 180,
+        720: 300
+    },
+    RATE: 1300
+}
+
+export {CONSTANTS};
