+++ src/modules/checkDayFilling.js
@@ -1,0 +1,11 @@
+const checkDayFilling = (ticketsData) => {
+    let filledDay = false;
+    for (const ticket in ticketsData) {
+        if (ticketsData[ticket] !== "") {
+            filledDay = true;
+        }
+    }
+    return filledDay;
+}
+
+export { checkDayFilling };
