+++ src/workers.js
@@ -1,0 +1,6 @@
+const WORKERS = {
+    Dasha: "Даша",
+    Nastya: "Настя"
+}
+
+export {WORKERS};
